Extract generic Table from Schedule page, make use of it, and rename Employees table
