+++ components/employees/schedule/ScheduleTableCell.tsx
@@ -10,7 +10,7 @@ const ScheduleTableCell: React.FC<IScheduleTableCell> = ({ value, onBlur }) => {
 
     return <input 
         className='w-10 border-2'
-        type="number" 
+        type='number'
         value={cellValue} 
         onChange={(e) => setCellValue(+e.target.value)}
         onBlur={() => onBlur(cellValue)}
