+++ interfaces/ActionsList.ts
@@ -1,0 +1,5 @@
+export interface IActionsList {
+    label: string
+    action: (() => void) | ((entityId: number) => void | Promise<void>)
+    textColor?: string
+}
