+++ interfaces/index.ts
@@ -1,4 +1,4 @@
 export * from './Menu'
 export * from './Schedule'
-export * from './Table'
+export * from './ActionsList'
 export * from './Dropdown'
